[apps-sdk] Refactor Apps SDK main entrypoint into modules
Split the monolithic Apps SDK server file into focused schemas, helper, events, widget, and REST modules while preserving endpoint contracts and MCP behavior to improve maintainability without functional changes.
